tv_control_agent/pyproject.toml
[project]
name = "tv_control_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "google-genai>=1.15.0",
    "python-dotenv>=1.1.0",
]

[tool.uv.sources]
google-adk = { git = "https://github.com/google/adk-python" }
